feat: atomic update config & split auth module into auth-data and auth-core
