+++ spring-cloud-alibaba-starters/spring-cloud-starter-alibaba-governance/governance-core/auth/auth-core/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,4 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+com.alibaba.cloud.governance.auth.XdsWebAutoConfiguration,\
+com.alibaba.cloud.governance.auth.XdsWebFluxAutoConfiguration,\
+com.alibaba.cloud.governance.auth.AuthValidatorAutoConfiguration
