+++ spring-cloud-alibaba-starters/spring-cloud-starter-alibaba-governance/governance-core/auth/auth-data/src/main/java/com/alibaba/cloud/governance/auth/rule/IdentityRule.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2013-2018 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.alibaba.cloud.governance.auth.rule;
+
+import com.alibaba.cloud.governance.common.matcher.StringMatcher;
+import com.alibaba.cloud.governance.common.rule.AndRule;
+
+public class IdentityRule {
+
+	private String name;
+
+	// in k8s, it is service account
+	private AndRule<StringMatcher> identities;
+
+	public IdentityRule(String name, AndRule<StringMatcher> identities) {
+		this.name = name;
+		this.identities = identities;
+	}
+
+	public AndRule<StringMatcher> getIdentities() {
+		return identities;
+	}
+
+	public String getName() {
+		return name;
+	}
+
+}
