--- spring-cloud-alibaba-starters/spring-cloud-starter-alibaba-governance/governance-core/auth/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-com.alibaba.cloud.governance.auth.XdsWebAutoConfiguration,\
-com.alibaba.cloud.governance.auth.XdsWebFluxAutoConfiguration
