+++ spring-cloud-alibaba-starters/spring-cloud-starter-alibaba-governance/resource-transform/istio-resource-transform/src/main/java/com/alibaba/cloud/governance/istio/PilotExchanger.java
@@ -47,7 +47,6 @@ public class PilotExchanger {
 
 	private void observeListeners(List<Listener> listeners) {
 		synchronized (ldsProtocol) {
-			ldsProtocol.clearCache();
 			ldsProtocol.resolveAuthRules(listeners);
 			Set<String> resourceName = ldsProtocol.getRouteNames(listeners);
 			if (resourceName != null && !resourceName.isEmpty()) {
