+++ spring-cloud-alibaba-starters/spring-cloud-starter-alibaba-governance/resource-transform/istio-resource-transform/src/main/java/com/alibaba/cloud/governance/istio/protocol/impl/CdsProtocol.java
@@ -49,11 +49,6 @@ public class CdsProtocol extends AbstractXdsProtocol<Cluster> {
 		return clusters;
 	}
 
-	@Override
-	protected void clearCache() {
-
-	}
-
 	public Set<String> getEndPointNames(List<Cluster> clusters) {
 		Set<String> endpoints = new HashSet<>();
 		for (Cluster cluster : clusters) {
