+++ spring-cloud-alibaba-starters/spring-cloud-starter-alibaba-governance/resource-transform/istio-resource-transform/src/main/java/com/alibaba/cloud/governance/istio/protocol/impl/EdsProtocol.java
@@ -48,11 +48,6 @@ public class EdsProtocol extends AbstractXdsProtocol<ClusterLoadAssignment> {
 	}
 
 	@Override
-	protected void clearCache() {
-
-	}
-
-	@Override
 	public String getTypeUrl() {
 		return "type.googleapis.com/envoy.config.endpoint.v3.ClusterLoadAssignment";
 	}
