+++ spring-cloud-alibaba-starters/spring-cloud-starter-alibaba-governance/resource-transform/istio-resource-transform/src/main/java/com/alibaba/cloud/governance/istio/protocol/impl/RdsProtocol.java
@@ -197,11 +197,6 @@ public class RdsProtocol extends AbstractXdsProtocol<RouteConfiguration> {
 	}
 
 	@Override
-	protected void clearCache() {
-
-	}
-
-	@Override
 	public String getTypeUrl() {
 		return "type.googleapis.com/envoy.config.route.v3.RouteConfiguration";
 	}
